api-archive/requirements.txt
google-auth 
google-auth-oauthlib 
requests